+++ freeciv/freeciv/common/unit.c
@@ -2381,10 +2381,10 @@ enum unit_upgrade_result unit_upgrade_info(const struct unit *punit,
   Returns the amount of movement points successfully performing the
   specified action will consume in the actor unit.
 **************************************************************************/
-int unit_pays_mp_for_action(const struct action *paction,
+long unit_pays_mp_for_action(const struct action *paction,
                             const struct unit *punit)
 {
-  int mpco;
+  long mpco;
 
   mpco = get_target_bonus_effects(NULL,
                                   unit_owner(punit),
